move items with tests to items.c
